requirements.txt
alabaster==0.7.12
amqp==2.3.2
appnope==0.1.0
argon2-cffi==18.3.0
arrow==0.12.1
atomicwrites==1.2.1
attrs==18.2.0
Babel==2.6.0
backcall==0.1.0
beautifulsoup4==4.6.3
billiard==3.5.0.4
binaryornot==0.4.4
celery==4.2.1
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
Click==7.0
cookiecutter==1.6.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.2
decorator==4.3.0
defusedxml==0.5.0
Django==2.0.9
django-allauth==0.38.0
django-cors-headers==2.4.0
django-coverage-plugin==1.6.0
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-environ==0.4.5
django-extensions==2.1.3
django-model-utils==3.1.2
django-redis==4.9.0
djangorestframework==3.9.0
docopt==0.6.2
docutils==0.14
factory-boy==2.11.1
Faker==1.0.0
flake8==3.6.0
flower==0.9.2
future==0.17.1
idna==2.7
imagesize==1.1.0
inflection==0.3.1
ipdb==0.11
ipython==7.1.1
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.13.1
Jinja2==2.10
jinja2-time==0.2.0
kombu==4.2.1
lxml==4.2.5
MarkupSafe==1.1.0
mccabe==0.6.1
more-itertools==4.3.0
multitasking==0.0.7
mypy==0.641
mypy-extensions==0.4.1
numpy==1.15.2
oauthlib==2.1.0
packaging==18.0
pandas==0.23.4
pandas-datareader==0.7.0
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
pipreqs==0.4.9
pluggy==0.8.0
poyo==0.4.2
prompt-toolkit==2.0.7
psycopg2==2.7.4
ptyprocess==0.6.0
py==1.7.0
pycodestyle==2.4.0
pycparser==2.19
pyflakes==2.0.0
Pygments==2.2.0
pyparsing==2.3.0
pytest==3.10.1
pytest-django==3.4.4
pytest-sugar==0.9.2
python-dateutil==2.7.3
python-decouple==3.1
python-slugify==1.2.6
python3-openid==3.1.0
pytz==2018.7
Quandl==3.4.2
redis==2.10.6
requests==2.19.1
requests-oauthlib==1.0.0
requirements==0.1
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.8.2
sphinx-bootstrap-theme==0.6.5
sphinx-sitemap==0.3.1
sphinxcontrib-plantuml==0.12
sphinxcontrib-websupport==1.1.0
sqlparse==0.2.4
termcolor==1.1.0
text-unidecode==1.2
tornado==5.1.1
traitlets==4.3.2
typed-ast==1.1.0
Unidecode==1.0.22
uritemplate==3.0.0
urllib3==1.23
vine==1.1.4
wcwidth==0.1.7
Werkzeug==0.14.1
whichcraft==0.5.2
wrapt==1.10.11
yarg==0.1.9
